Expand template domain definition
